+++ NEWS.md
@@ -1,6 +1,26 @@
 tag next
 --------------------------------
 
+tag k0.8.2
+--------------------------------
+- ffmpeg: fix some bug
+- ijkio: update and modify features
+- ijkplayer: support don't calculate real frame rate, the first frame will speed up
+
+tag k0.8.1
+--------------------------------
+- ffmpeg: support dns cache
+- ijkio: support inject extra node
+
+tag k0.8.0
+--------------------------------
+- ffmpeg: upgrade to 3.3
+- ffmpeg: enable flac
+- android: support sync mediacodec
+- android: support framedrop when use mediacodec
+- openssl: upgrade to 1.0.2k
+- jni4android: upgrade to v0.0.2
+
 tag k0.7.9
 --------------------------------
 - ffmpeg: add tcp timeout control